requirements.txt
azure-identity==1.25.2
azure-identity-broker==1.3.0 # Token protection policy requires proof-of-possession (PoP) tokens, which needs the WAM broker
pywin32==311 ; sys_platform == "win32"
requests==2.32.5
markitdown[docx]==0.1.5
